Update instructions for setting up python env
Rather than relying on conda for sphinx and sphinx_rtd_theme, using pip to
install everything in our requirements file makes sure every developer is using
the same versions of all the sphinx-related packages. Most importantly, it lets
everyone have access to the branch off of sphinx_rtd_theme 0.2.5b1 that allows
the version dropdown menu.

Updated the instructions on "working on the docs" page to account for this, and
also added link back to homepage in sidebar (above contents listing the rest of
the sections)